requirements.txt
numpy
scipy
pillow
pandas
tqdm
matplotlib
pytest
prospector
sphinx
pyinstaller==6.11.1